revert clientoptions
