--- cli/runtime/config/cli_unstable_version_selector_test.go
@@ -1,1 +1,0 @@
-package config
